moving ValidationExceptions to correct package
