+++ src/main/java/de/konfidas/ttc/exceptions/LogMessageValidationException.java
@@ -1,11 +1,11 @@
-package de.konfidas.ttc.validation;
+package de.konfidas.ttc.exceptions;
 
 import de.konfidas.ttc.messages.LogMessage;
 
 public abstract class LogMessageValidationException extends ValidationException{
     LogMessage msg;
 
-    LogMessageValidationException(LogMessage msg){
+    public LogMessageValidationException(LogMessage msg){
         super("Validating Log Message "+msg.getFileName()+" failed",null);
         this.msg = msg;
     }
