+++ src/main/java/de/konfidas/ttc/validation/AggregatedValidator.java
@@ -1,5 +1,6 @@
 package de.konfidas.ttc.validation;
 
+import de.konfidas.ttc.exceptions.ValidationException;
 import de.konfidas.ttc.tars.LogMessageArchive;
 
 import java.util.ArrayList;
